glfw: fix crash when window size is <= 0
Windows sets the window size to (0,0) when minimizing the window. That leads to eventual division by zero errors and opengl does not want to attach a framebuffer with that size, resulting in it being incomplete. After maximizing again, it tries to unload that framebuffer, but it was never loaded. Minosoft crashes with an invalid framebuffer state: deleted