font texture: properly copy alpha channel to rgb components
This finally fixes all font rendering issues with compression.

Previously the font textures were in the aaaa format, since 1.20 they seem to be in the 111a format (which is completely valid and reduces file size). the problem is, that the with font compression just the red channel will be used. This one is always 1, so all chars are are quare on screen. Not the alpha channel is force copied to the rgb channels