chunk manager: calculate light async
The chunk renderer set the neighbours for the chunk (locked) and calculated the light during that time. That was really causing lag stuttering and took way too long.

Now it is async, but there is still some smell, the light is calculated 8x instead of just once. Not sure why. To be investigated.